requirements.txt
numpy
torch
torchvision
argparse
matplotlib
pandas
pillow
tqdm